Se agregan servicios para generar nota del examen

+++ back/src/main/java/pep3/savingsemester/back/Models/RespuestaM.java
@@ -7,5 +7,5 @@ import lombok.*;
 @Data
 public class RespuestaM {
     public Integer idPregunta;
-    public Integer respuesta;
+    public Double respuesta;
 }
